+++ src/models/iIdHolder.ts
@@ -1,0 +1,8 @@
+import { LogLevel } from './logLevel';
+
+export interface iIdHolder {
+  readonly id: string;
+  readonly customName: string;
+
+  log(level: LogLevel, message: string): void;
+}