+++ src/models/index.ts
@@ -8,5 +8,6 @@ export * from './deviceConfig';
 export * from './excessEnergyConsumerSettings';
 export * from './iTemperaturDataPoint';
 export * from './logLevel';
+export * from './objectSettings';
 export * from './temperatureSettings';
 export * from './timeCallback';