+++ src/server/devices/baseDeviceInterfaces/iBaseDevice.ts
@@ -1,17 +1,14 @@
-import { DeviceSettings, LogLevel } from '../../../models';
+import { DeviceSettings } from '../../../models';
 import { DeviceType } from '../deviceType';
 import { DeviceInfo } from '../DeviceInfo';
 import { DeviceCapability } from '../DeviceCapability';
+import { iIdHolder } from '../../../models/iIdHolder';
 
-export interface iBaseDevice {
+export interface iBaseDevice extends iIdHolder {
   readonly settings: DeviceSettings | undefined;
   deviceType: DeviceType;
   info: DeviceInfo;
   readonly deviceCapabilities: DeviceCapability[];
-
-  readonly id: string;
-
-  log(level: LogLevel, message: string): void;
 
   persistDeviceInfo(): void;