+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Generalize device Settings
+
 ## 1.3.14-alpha.0 (2022-11-16)
 
 * (theimo1221) Log Forced Device Queries